+++ Tapjoy/src/main/java/com/mopub/mobileads/TapjoyRewardedVideo.java
@@ -17,12 +17,14 @@ import com.tapjoy.TJConnectListener;
 import com.tapjoy.TJError;
 import com.tapjoy.TJPlacement;
 import com.tapjoy.TJPlacementListener;
+import com.tapjoy.TJPlacementVideoListener;
 import com.tapjoy.TJVideoListener;
 import com.tapjoy.Tapjoy;
 import com.tapjoy.TapjoyLog;
 
 import org.json.JSONException;
 
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOULD_REWARD;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_ATTEMPTED;
@@ -168,7 +170,7 @@ public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
                     MoPubLog.log(CUSTOM, ADAPTER_NAME, "Unable to parse auction data.");
                 }
             }
-
+            tjPlacement.setVideoListener(sTapjoyListener);
             tjPlacement.requestContent();
             MoPubLog.log(placementName, LOAD_ATTEMPTED, ADAPTER_NAME);
         } else {
@@ -241,7 +243,7 @@ public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
         }
     }
 
-    private static class TapjoyRewardedVideoListener implements TJPlacementListener, CustomEventRewardedVideoListener, TJVideoListener {
+    private static class TapjoyRewardedVideoListener implements TJPlacementListener, CustomEventRewardedVideoListener, TJPlacementVideoListener {
         @Override
         public void onRequestSuccess(TJPlacement placement) {
             if (!placement.isContentAvailable()) {
@@ -264,15 +266,19 @@ public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
 
         @Override
         public void onContentShow(TJPlacement placement) {
-            Tapjoy.setVideoListener(this);
             MoPubRewardedVideoManager.onRewardedVideoStarted(TapjoyRewardedVideo.class, TAPJOY_AD_NETWORK_CONSTANT);
             MoPubLog.log(SHOW_SUCCESS, ADAPTER_NAME);
         }
 
         @Override
         public void onContentDismiss(TJPlacement placement) {
-            Tapjoy.setVideoListener(null);
             MoPubRewardedVideoManager.onRewardedVideoClosed(TapjoyRewardedVideo.class, TAPJOY_AD_NETWORK_CONSTANT);
+        }
+
+        @Override
+        public void onClick(TJPlacement placement) {
+            MoPubLog.log(CLICKED, ADAPTER_NAME);
+            MoPubRewardedVideoManager.onRewardedVideoClicked(TapjoyRewardedVideo.class, TAPJOY_AD_NETWORK_CONSTANT);
         }
 
         @Override
@@ -286,19 +292,25 @@ public class TapjoyRewardedVideo extends CustomEventRewardedVideo {
         }
 
         @Override
-        public void onVideoStart() {
+        public void onVideoStart(TJPlacement tjPlacement) {
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Tapjoy rewarded video started for placement " +
+                    tjPlacement + ".");
 
         }
 
         @Override
-        public void onVideoError(int statusCode) {
+        public void onVideoError(TJPlacement tjPlacement, String message) {
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Tapjoy rewarded video failed for placement " +
+                    tjPlacement + "with error" + message);
+
         }
 
         @Override
-        public void onVideoComplete() {
+        public void onVideoComplete(TJPlacement tjPlacement) {
             MoPubLog.log(CUSTOM, ADAPTER_NAME, "Tapjoy rewarded video completed");
             MoPubRewardedVideoManager.onRewardedVideoCompleted(TapjoyRewardedVideo.class, TAPJOY_AD_NETWORK_CONSTANT, MoPubReward.success(MoPubReward.NO_REWARD_LABEL, MoPubReward.NO_REWARD_AMOUNT));
             MoPubLog.log(SHOULD_REWARD, ADAPTER_NAME, MoPubReward.NO_REWARD_AMOUNT, MoPubReward.NO_REWARD_LABEL);
+
         }
     }